refactor: data template
